refactor: update dj package to use new name
